+++ src/involvement/forms/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from .form import ReferenceFormSet
+from .application_form import ApplicationForm
+from .approval_form import ApprovalForm
+from .appointment_form import AppointmentForm
